refactor : testcode 메소드별로 파일 분리하기
